Add Socket Inputter
This commit adds a SocketInputter that opens a Unix Socket and reads messages from that socket.
By virtue of only being able to read bytes, we also add in a Parser thing that is used by byte stream inputters
to parse individual messages out of the byte stream. This commit contains two Parsers - a new line parser
that parses messages by line, and a JSON one that parses things as a JSON object