Move to Pool for message batches
This commit moves to using a sync.Pool to manage message batches. This
allows us to reuse batches rather than cycle them quickly (as is the
case for our inputs at the moment that only every produce one message
batches). This allows us to amortise the cost of the allocs for batches
over the length of the input