Add Link type
This commit moves us away from pipes between elements in the pipeline
being basic strings, to having a bit more structure. In particular, we
introduce a `LinkType` that will allow control over how the link
functions. We have two Link Types at the moment - Normal, which is how
links currently work, and Buffer which is for buffer connections that
will be implemented after this